refactor(consult): split the 4.3k-line consult.rs into a directory module
`consult.rs` was one of the two "swollen organs" the 2026-07-02 health
review named. Split it into `src/consult/` with three submodules along the
seams that review identified, `mod.rs` re-exporting the public surface so
every `crate::consult::X` call site is unchanged:

- **shaping** — the provider-drift / request-shaping knowledge (`ModelShape`,
  `ModelCaps`, the id-classifiers, `thinking_params`/`request_params`, the
  effort/budget consts). The part that rots as providers ship models.
- **prompts** — the preambles and framers (`report`/`consult`/`oneshot`/
  `batch_preamble`, `phase_preamble`/`Phase`, `consult_user_prompt`,
  `deliberation_prompt`, `PromptOverrides`). The text a model reads.
- **engine** — the tool-loop machinery and orchestration (`Arm`/`PhaseRunner`/
  `run_phase`, the view-image break-rewrite, finalize/resume, `RunExplore`,
  `consult`/`oneshot`/`explore_with`/`deliberate_direct`, `ConsultConfig`).

Dependency direction is clean: engine depends on prompts and shaping, never
the reverse. `ConsultAttachment` lives with prompts (not engine) because
`consult_user_prompt` takes it — putting it in engine would invert that edge.

Pure behavior-preserving move: no logic, string, preamble, const, or test
assertion changed; only per-file imports and the mod.rs glue are new. The
46 consult unit tests were distributed to the submodule each exercises
(shaping 1, prompts 9, engine 36). Full suite green (308 lib + all
integration), clippy clean (the one warning is pre-existing, in telemetry.rs).

Internal refactor — no user-facing change, so no CHANGELOG entry.
