+++ src/consult/engine.rs
@@ -1,23 +1,4 @@
-//! `consult` and the seams it's composed from, across providers.
-//!
-//! One primitive — [`run_phase`]: a model + preamble + an injected toolset, run as
-//! a bounded tool loop. Every tool on the surface is that loop wearing different
-//! clothes:
-//!
-//! - [`consult`] — a capable model · `{run_kaish, explore′}` · optional context → a
-//!   cited answer. No rigid explorer→synth hand-off: the capable model decides when
-//!   to delegate a broad sweep to the cheap [`RunExplore`] sub-agent vs. read a span
-//!   directly. The `explore′` sweep ([`report_preamble`]) is internal to this loop.
-//! - [`oneshot`] — a capable model · no tools · the caller's context → a direct
-//!   answer. The thin counterpart: one upstream request, no codebase access.
-//!
-//! Each phase arrives as a resolved [`Arm`]: its own client (type-erased — the
-//! decided plumbing fork, `docs/casts.md`), model, request params, and caps. The
-//! server resolves a cast's slots into arms ([`Arm::from_slot`]); a cast whose
-//! explorer and synth live on different backends — different wire protocols,
-//! even — runs each phase on its own client through the same loop primitive.
-//! Each tool gets its own fresh [`KaishWorker`] (a kernel rooted at the
-//! project), and so does every `explore′` delegation.
+//! The consult tool-loop engine and orchestration.
 
 use std::collections::HashSet;
 use std::future::Future;
@@ -44,623 +25,17 @@ use crate::attach::Attachment;
 use crate::config::{Backend, Defaults, ModelRole, ModelSlot};
 use crate::credentials::ProviderKind;
 use crate::explorer::RunKaish;
-use crate::kaish_syntax::kaish_syntax_core;
 use crate::progress::{NullSink, PhaseEvent, ProgressSink};
 use crate::sandbox::{KaishWorker, SandboxConfig};
 use crate::session::{QaTurn, SessionStore};
 use crate::tool_span::traced;
 use crate::view_image::ViewImage;
 
-/// Splice the operator's house rules (if any) onto a phase preamble. The base
-/// preamble functions stay pure (and their tests byte-for-byte stable); this is
-/// the one seam that folds in the assembled `[context]` block. Every phase that
-/// drives a model uses it — the `consult` driver, the toolless `oneshot`, *and* the
-/// nested `explore′` sweep — so the explorer orients on the same guidance the driver
-/// does (it helps *search*, not just the answer).
-/// `None` returns the base unchanged: a server with no `[context]` files runs
-/// exactly the historical preamble.
-///
-/// Framed as standing background, not the question, and positively (per the
-/// `positive-prompt-framing` discipline): tell the model what the block *is* and
-/// how to use it — conventions to honor while investigating — rather than fencing
-/// it off. It sits *after* the base so the tool's own role framing leads.
-fn with_house_rules(base: String, house_rules: Option<&str>) -> String {
-    match house_rules {
-        None => base,
-        Some(rules) => format!(
-            "{base}\n\n\
-             --- Operator house rules for this codebase ---\n\
-             The agent you're helping configured the guidance below — project \
-             conventions and working preferences for this repository. Treat it as \
-             trusted standing context: honor it as you investigate and when you write \
-             your answer. It's background about how this codebase works, not the \
-             question you're answering.\n\n{rules}"
-        ),
-    }
-}
-
-/// Operator preamble (system-prompt) overrides per phase, from the `[prompts]`
-/// config table. `None` for a phase means "use the built-in" — so an empty table
-/// is byte-for-byte the historical preambles. **Full replace** by decision: an
-/// override *is* the role framing, verbatim; the kaish operating contract is not
-/// re-appended here because it independently rides the `run_kaish` tool
-/// description (`run_kaish_tool_description`), so the model keeps the shell
-/// contract even when an operator rewrites the prose. Empty/whitespace values are
-/// refused at config load (`config.rs::merge_prompts`) — a blank system prompt is
-/// never the intent. House rules still append on top (see [`phase_preamble`]):
-/// `[prompts]` replaces the *role* framing, `[context]` adds *project* guidance —
-/// orthogonal axes.
-#[derive(Debug, Clone, Default)]
-pub struct PromptOverrides {
-    /// Replaces [`report_preamble`] — the nested `explore′` sweep inside `consult`.
-    pub explorer: Option<String>,
-    /// Replaces [`consult_preamble`] — the `consult` driver.
-    pub consult: Option<String>,
-    /// Replaces [`oneshot_preamble`] — the thin, toolless `oneshot`.
-    pub oneshot: Option<String>,
-    /// Replaces [`batch_preamble`] — the offline, max-thinking `batch_submit`. A key
-    /// of its own (not shared with `oneshot`) because the batch lane is a different
-    /// behavioral contract: one response, no follow-up, spend on depth.
-    pub batch: Option<String>,
-}
-
-/// Resolve one phase's full system prompt: the operator override if set, else the
-/// built-in `default`, then the static repo `orientation` map, then house rules.
-/// The single composition point for every model-driven phase, so override +
-/// `[orientation]` + `[context]` layering is identical everywhere. Order: role
-/// framing → the file map (immediately useful context) → operator house rules.
-fn phase_preamble(
-    override_: Option<&str>,
-    default: fn() -> String,
-    orientation: Option<&str>,
-    house_rules: Option<&str>,
-) -> String {
-    let mut base = override_.map(str::to_string).unwrap_or_else(default);
-    if let Some(map) = orientation {
-        base.push_str("\n\n");
-        base.push_str(map); // carries its own `PROJECT FILES.` header
-    }
-    with_house_rules(base, house_rules)
-}
-
-/// The model-driven phases whose system prompt kaibo composes. One enum so the three
-/// per-phase decisions — *which* built-in default, *which* `[prompts]` override key,
-/// and *whether* the phase reads the project (so the `[orientation]` map + `[context]`
-/// house rules splice) — live in exactly one place, [`resolve_phase_preamble`]. Every
-/// live tool routes through it, and so does the `kaibo://prompts` resource, so what the
-/// resource shows can never drift from what a call actually sends the model.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum Phase {
-    /// The explorer sweep: `explore`, the nested `explore′` inside `consult`, and
-    /// `deliberate`'s dossier phase all share this one.
-    Explorer,
-    /// The `consult` driver.
-    Consult,
-    /// The thin, toolless `oneshot`.
-    Oneshot,
-    /// The offline synth: `batch_submit` and `deliberate`'s synth, on either lane.
-    Batch,
-}
-
-impl Phase {
-    /// The four phases, for callers that enumerate every prompt (the resource).
-    pub const ALL: [Phase; 4] = [
-        Phase::Explorer,
-        Phase::Consult,
-        Phase::Oneshot,
-        Phase::Batch,
-    ];
-
-    /// A short, stable label for a phase — the resource header and the tools it drives.
-    pub fn label(self) -> &'static str {
-        match self {
-            Phase::Explorer => "explorer (explore · consult sweep · deliberate dossier)",
-            Phase::Consult => "consult driver",
-            Phase::Oneshot => "oneshot",
-            Phase::Batch => "batch / deliberate synth (offline)",
-        }
-    }
-
-    /// The built-in default preamble for this phase.
-    fn default_preamble(self) -> fn() -> String {
-        match self {
-            Phase::Explorer => report_preamble,
-            Phase::Consult => consult_preamble,
-            Phase::Oneshot => oneshot_preamble,
-            Phase::Batch => batch_preamble,
-        }
-    }
-
-    /// This phase's `[prompts]` override key (per-slot preamble already folded in
-    /// upstream). Public so the `kaibo://prompts` resource can report which phases carry
-    /// an active override without re-encoding the phase→key mapping.
-    pub fn override_in(self, p: &PromptOverrides) -> Option<&str> {
-        match self {
-            Phase::Explorer => p.explorer.as_deref(),
-            Phase::Consult => p.consult.as_deref(),
-            Phase::Oneshot => p.oneshot.as_deref(),
-            Phase::Batch => p.batch.as_deref(),
-        }
-    }
-
-    /// Does this phase read the project? The explorer sweep and the `consult` driver
-    /// do — so they get the `[orientation]` map and `[context]` house rules spliced.
-    /// `oneshot` and the offline `batch` synth own their context (the caller supplies
-    /// it), so neither project layer reaches them — the seam that used to sit as a bare
-    /// `None, None` at each of those call sites now lives here, in one place.
-    pub fn reads_project(self) -> bool {
-        matches!(self, Phase::Explorer | Phase::Consult)
-    }
-
-    /// Which cast slot's `preamble` frames this phase: the **explorer** slot drives the
-    /// explorer sweep; the **synth** slot drives every synth phase (`consult`, `oneshot`,
-    /// and the offline `batch`/`deliberate` synth). Lets the `kaibo://prompts/{cast}`
-    /// resource attribute a phase's framing to the slot that set it — the same slot→phase
-    /// mapping [`crate::config::Cast::resolved_prompts`] applies.
-    pub fn slot_role(self) -> ModelRole {
-        match self {
-            Phase::Explorer => ModelRole::Explorer,
-            Phase::Consult | Phase::Oneshot | Phase::Batch => ModelRole::Synth,
-        }
-    }
-}
-
-/// Compose one phase's full system prompt through the single layering point. Picks the
-/// operator override (else the built-in) for `phase`, then — for the project-reading
-/// phases only — splices the `[orientation]` map and `[context]` house rules. This is
-/// what every live tool builds and what the `kaibo://prompts` resource renders, so the
-/// resource is exactly the code path, not a restatement of it.
-pub fn resolve_phase_preamble(
-    phase: Phase,
-    prompts: &PromptOverrides,
-    orientation: Option<&str>,
-    house_rules: Option<&str>,
-) -> String {
-    // The phase decides whether the project layers apply — pass them unconditionally
-    // and let `reads_project` gate, so no call site re-encodes that rule.
-    let (orientation, house_rules) = if phase.reads_project() {
-        (orientation, house_rules)
-    } else {
-        (None, None)
-    };
-    phase_preamble(
-        phase.override_in(prompts),
-        phase.default_preamble(),
-        orientation,
-        house_rules,
-    )
-}
-
-/// Explorer preamble: gather and organize evidence, don't conclude. Composes the
-/// shared [`kaish_syntax_core`] so the shell idioms and exit-code contract are
-/// stated in exactly one place.
-pub fn report_preamble() -> String {
-    let core = kaish_syntax_core();
-    format!(
-        "You are a code explorer. You build a complete, accurate picture of the code \
-         a question touches and hand it to a synthesizer who writes the final \
-         answer — so your work is to gather grounded evidence and cite it exactly. \
-         {core}\n\n\
-         HOW TO READ. Read for the whole picture in as few looks as possible — the \
-         context window is yours to fill, so read in wide passes. A short file: read \
-         it WHOLE with `cat -n FILE` — one read hands you its imports, its context, \
-         and exact line numbers together. A big file (`wc -l FILE` if unsure): walk it \
-         in wide spans of a few hundred lines — `cat -n FILE | sed -n '1,400p'`, then \
-         `'401,800p'`, then `'801,1200p'` — so each look lands a whole run of related \
-         code together: a type with its impl, a function with the code around its call \
-         sites, an import block with what uses it. If a whole-file read comes back \
-         truncated (exit 3, a head+tail sample), it was too big for one look — walk it \
-         in those wide spans. To locate something across files, take the surrounding \
-         context in the same call — `grep -rn -B4 -A8 PATTERN .` returns each match \
-         with the lines around it, ready to understand.\n\n\
-         HOW TO INVESTIGATE. Aim for the complete set of relevant locations. Follow \
-         each key symbol to where it is defined and where it is used; chase anything \
-         that puzzles you until it is clear — a confusing spot usually hides the \
-         thing you need. Follow each thread while you are already in the code, so one \
-         thorough pass leaves you the complete picture.\n\n\
-         WHAT TO PRODUCE. A curated report for the synthesizer, in these sections:\n\
-         - SummaryOfFindings: what you concluded, in a few sentences.\n\
-         - RelevantLocations: for each location that matters — the concrete \
-         `file:line`, the key symbols there (functions, types, fields), a short \
-         verbatim snippet, and what it means for the question.\n\
-         - ExplorationTrace: the path you took, when it helps the synthesizer trust \
-         the result.\n\
-         Keep it tight and evidence-first. The synthesizer trusts your citations and \
-         builds on them, so ground every claim in an exact `file:line` — that \
-         exactness is the whole value of your report."
-    )
-}
-
-/// Token budget for model "thinking"/reasoning, for the providers that expose a
-/// request-time toggle. Sized well under [`ConsultConfig`]'s `max_tokens` so the
-/// reasoning never starves the actual answer (a thinking model that spends its
-/// whole budget reasoning returns empty content — we saw exactly that on Gemma).
-/// Anthropic additionally *requires* `max_tokens > budget_tokens`.
-pub const THINKING_BUDGET: u64 = 8192;
-
-/// Does this Gemini id belong to the *pure* 3-line (e.g. `gemini-3-pro-preview`),
-/// which takes `thinkingLevel` rather than the 2.5-era `thinkingBudget`?
-///
-/// The boundary is **empirical, not nominal**: `gemini-3.5-flash` *accepted*
-/// `thinkingBudget` in the 2026-06-06 live test, so switching it to `thinkingLevel`
-/// would be a silent regression of a working default. We only flip the ids the
-/// official API + gemini-cli confirm want a level — `gemini-3-…` — and leave the
-/// `3.5` minor line (and 2.x) on budget. Any new id past these wants a live probe,
-/// not a guess. See `docs/issues.md` "Per-model request shaping".
-fn is_gemini3_level(model: &str) -> bool {
-    model == "gemini-3" || model.starts_with("gemini-3-")
-}
-
-/// The per-role thinking-depth lever for the models that expose one as a request
-/// param (Anthropic adaptive's `output_config.effort`, DeepSeek's `reasoning_effort`).
-/// A passthrough string the provider validates — like a model id — so a new level
-/// lands without a code change. Default for both roles unless a slot or the
-/// per-role `[defaults]` tunes it.
-pub const DEFAULT_EFFORT: &str = "high";
-
-/// Which Anthropic models want **adaptive** thinking (`{type:"adaptive"}` plus an
-/// `output_config.effort`) instead of the legacy `{type:"enabled", budget_tokens}`.
-///
-/// **Empirical — confirm by probe** (the discipline of [`is_gemini3_level`]): Opus
-/// 4.7/4.8 and Fable 5 *reject* enabled/budget and sampling outright (400); Opus 4.6 /
-/// Sonnet 4.6 take adaptive too — it's the recommended shape, `budget_tokens` is
-/// deprecated there. Everything older, and Haiku 4.5, stays on enabled/budget. Matched
-/// by `contains` (not `starts_with`, unlike `is_gemini3_level`) so a vendor-prefixed id
-/// still resolves. Add ids as they ship; a slot (or `[defaults]`) can force a tier
-/// via `thinking_style`.
-fn is_anthropic_adaptive(model: &str) -> bool {
-    ["opus-4-6", "opus-4-7", "opus-4-8", "sonnet-4-6", "fable-5"]
-        .iter()
-        .any(|tier| model.contains(tier))
-}
-
-/// What one (provider, model) can perceive — and *how* an image reaches it. Capability
-/// data on the same seam as [`ModelShape`], resolved per model slot: an explicit config
-/// override wins, else the built-in classifier. Toolsets are assembled from resolved
-/// caps (a vision model gets `view_image` when vision-in lands; a blind model never
-/// sees the tool), so a capability mismatch is structural, not a runtime surprise.
-///
-/// The real predicate `view_image` rides on is **see ∧ transport**: a model can *see*
-/// (`vision`) AND the chosen channel can *carry* the image. Anthropic/Gemini carry an
-/// image inside a tool result (`tool_result_images`); OpenAI's wire forbids it (rig
-/// 400s before sending), so an OpenAI VLM must receive the image on the user-turn
-/// channel instead — the break-rewrite-resume path in [`run_phase`]. The two bools let
-/// the toolset gate `view_image` on `vision` while the loop gates the rewrite on
-/// `vision && !tool_result_images`.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct ModelCaps {
-    /// Accepts image parts in model context (vision-in).
-    pub vision: bool,
-    /// Does this model's *transport* carry an image inside a tool result? When false,
-    /// a seen image must ride the user-turn channel instead (see [`Arm::rewrites_view_image`]).
-    pub tool_result_images: bool,
-}
-
-impl ModelCaps {
-    /// Resolve the caps for `model` under `kind`. `vision_override` is the per-slot
-    /// config escape hatch (`vision = true/false` in the role table) — it pins the
-    /// vision answer in both directions; `None` asks the classifier. The transport
-    /// channel is a property of the wire protocol alone (`kind`), never overridden.
-    pub fn resolve(kind: ProviderKind, model: &str, vision_override: Option<bool>) -> Self {
-        let vision = vision_override.unwrap_or_else(|| is_vision_capable(kind, model));
-        Self {
-            vision,
-            tool_result_images: transport_supports_tool_result_images(kind),
-        }
-    }
-}
-
-/// Does this wire protocol carry an image *inside a tool result*? Anthropic
-/// (`tool_result` image blocks) and Gemini (`functionResponse` inline data) do — it's
-/// documented and first-class. The OpenAI wire forbids images in a `role:tool` message
-/// and rig enforces it before sending (`ToolResultContent::Image(_) => Err(..)` in
-/// `openai/completion/mod.rs`), so a `view_image` result there must instead be
-/// delivered as an `image_url` part on a **user** turn. DeepSeek is moot — vision-blind,
-/// so `view_image` never attaches. Branch the rewrite on *this*, not on `kind == Openai`:
-/// the next no-tool-result-image provider is a table entry, not a new `if`.
-fn transport_supports_tool_result_images(kind: ProviderKind) -> bool {
-    match kind {
-        ProviderKind::Anthropic | ProviderKind::Gemini => true,
-        ProviderKind::Openai => false,
-        // Vision-blind on the wire; the value is unreached (no view_image attaches),
-        // but "no tool-result image channel" is the honest answer.
-        ProviderKind::DeepSeek => false,
-    }
-}
-
-/// The built-in vision classifier. **Empirical — confirm by probe** (the discipline
-/// of [`is_anthropic_adaptive`]): boundaries reflect what the providers actually
-/// serve today, and a wrong guess fails loud at the provider, not silent here.
-fn is_vision_capable(kind: ProviderKind, _model: &str) -> bool {
-    match kind {
-        // Every current Claude completion id is multimodal-in (vision shipped with
-        // Claude 3; no text-only ids remain in the lineup).
-        ProviderKind::Anthropic => true,
-        // The gemini-* completion line is natively multimodal across 2.x/3.x.
-        ProviderKind::Gemini => true,
-        // DeepSeek chat/reasoner are text-only on the wire (docs/issues.md, media
-        // spine): images attached to a blind model must fail loud, not get dropped.
-        ProviderKind::DeepSeek => false,
-        // A generic OpenAI-compatible endpoint can front anything; vision is opt-in
-        // per slot (`vision = true` in the role table) rather than guessed from an
-        // arbitrary id.
-        ProviderKind::Openai => false,
-    }
-}
-
-/// How a given (provider, model) expresses "think" on the wire.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum ThinkingStyle {
-    /// Anthropic legacy: `{thinking:{type:"enabled",budget_tokens:N}}`.
-    AnthropicBudget,
-    /// Anthropic 4.6+: `{thinking:{type:"adaptive"}}` + `output_config.effort`.
-    AnthropicAdaptive,
-    /// Gemini 3-line: `generationConfig.thinkingConfig.thinkingLevel`.
-    GeminiLevel,
-    /// Gemini 2.5/3.5: `generationConfig.thinkingConfig.thinkingBudget`.
-    GeminiBudget,
-    /// DeepSeek V4 hybrids: `{thinking:{type:"enabled"}, reasoning_effort:<role>}`.
-    DeepSeekEffort,
-    /// No request-time toggle (the generic OpenAI path).
-    None,
-}
-
-/// Where this provider's wire format puts sampling (`temperature`/`top_p`).
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-enum SamplingPlacement {
-    /// Gemini nests it under `generationConfig` (camelCase `topP`).
-    GeminiGenerationConfig,
-    /// Anthropic/DeepSeek/OpenAI take it top-level (rig flattens into the body).
-    TopLevel,
-}
-
-/// Force a model's thinking style, overriding the built-in classifier. `Auto` (the
-/// default) classifies from the model id; the others pin a tier — the escape hatch for
-/// a new or misclassified Anthropic model (see `docs/config.md`). Carried on a cast
-/// slot (falling back to `[defaults]`) and resolved per arm via [`Arm::from_slot`].
-#[derive(Debug, Clone, Copy, PartialEq, Eq, Default)]
-pub enum ThinkingStyleOverride {
-    #[default]
-    Auto,
-    Adaptive,
-    Budget,
-}
-
-impl std::str::FromStr for ThinkingStyleOverride {
-    type Err = anyhow::Error;
-    fn from_str(s: &str) -> Result<Self> {
-        match s.trim().to_ascii_lowercase().as_str() {
-            "auto" => Ok(Self::Auto),
-            "adaptive" => Ok(Self::Adaptive),
-            "budget" => Ok(Self::Budget),
-            other => Err(anyhow!(
-                "thinking_style {other:?} is not one of auto|adaptive|budget"
-            )),
-        }
-    }
-}
-
-/// The request shape one (provider, model) wants — the unified, per-model home for
-/// request tuning. [`resolve`](ModelShape::resolve) classifies it (honoring an
-/// override); [`to_params`](ModelShape::to_params) emits the `additional_params` blob
-/// with the per-phase budget, sampling, and effort. This is what lets a cast whose
-/// explorer and synth straddle a capability line (a budget-tier Haiku explorer with an
-/// adaptive Sonnet 4.6 synth, say) fit each arm correctly.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub struct ModelShape {
-    thinking: ThinkingStyle,
-    /// Does this model accept sampling *while thinking is on*? Anthropic 400s on a
-    /// custom temperature under thinking (any tier — "temperature may only be set to 1
-    /// when thinking is enabled"; 4.7/4.8/Fable reject it outright), so `false` there;
-    /// DeepSeek/Gemini/OpenAI accept sampling alongside reasoning.
-    sampling_under_thinking: bool,
-    sampling_placement: SamplingPlacement,
-}
-
-impl ModelShape {
-    /// Resolve the shape for `model` under `kind`, honoring an explicit override.
-    pub fn resolve(kind: ProviderKind, model: &str, ovr: ThinkingStyleOverride) -> Self {
-        let thinking = match kind {
-            ProviderKind::Anthropic => {
-                let adaptive = match ovr {
-                    ThinkingStyleOverride::Auto => is_anthropic_adaptive(model),
-                    ThinkingStyleOverride::Adaptive => true,
-                    ThinkingStyleOverride::Budget => false,
-                };
-                if adaptive {
-                    ThinkingStyle::AnthropicAdaptive
-                } else {
-                    ThinkingStyle::AnthropicBudget
-                }
-            }
-            ProviderKind::Gemini => {
-                if is_gemini3_level(model) {
-                    ThinkingStyle::GeminiLevel
-                } else {
-                    ThinkingStyle::GeminiBudget
-                }
-            }
-            ProviderKind::DeepSeek => ThinkingStyle::DeepSeekEffort,
-            ProviderKind::Openai => ThinkingStyle::None,
-        };
-        let (sampling_under_thinking, sampling_placement) = match kind {
-            ProviderKind::Anthropic => (false, SamplingPlacement::TopLevel),
-            ProviderKind::Gemini => (true, SamplingPlacement::GeminiGenerationConfig),
-            ProviderKind::DeepSeek | ProviderKind::Openai => (true, SamplingPlacement::TopLevel),
-        };
-        Self {
-            thinking,
-            sampling_under_thinking,
-            sampling_placement,
-        }
-    }
-
-    /// Does this shape have a sink for a `thinking_budget` tunable? Only the two
-    /// explicit-budget styles emit it (`budget_tokens` / `thinkingBudget`); the
-    /// effort-driven styles (Anthropic adaptive, the Gemini 3-line, DeepSeek) and the
-    /// toggle-less openai path ignore a budget entirely. Lets the `kaibo://config`
-    /// render flag a per-slot `thinking_budget` that will never leave the process.
-    pub fn sinks_thinking_budget(&self) -> bool {
-        matches!(
-            self.thinking,
-            ThinkingStyle::AnthropicBudget | ThinkingStyle::GeminiBudget
-        )
-    }
-
-    /// Does this shape have a sink for an `effort` tunable? Only the effort-driven
-    /// styles route it (`output_config.effort` / `thinkingLevel` / `reasoning_effort`);
-    /// the budget styles and the openai path drop it. Counterpart to
-    /// [`sinks_thinking_budget`](Self::sinks_thinking_budget) for the render's no-op flag.
-    pub fn sinks_effort(&self) -> bool {
-        matches!(
-            self.thinking,
-            ThinkingStyle::AnthropicAdaptive
-                | ThinkingStyle::GeminiLevel
-                | ThinkingStyle::DeepSeekEffort
-        )
-    }
-
-    /// Does this shape actually send sampling (`temperature`/`top_p`)? Kaibo runs
-    /// thinking on by default, and a model that rejects sampling under thinking
-    /// (Anthropic, any tier) has it dropped — so a per-slot `temperature` there is
-    /// inert. The toggle-less openai path (`thinking == None`) keeps sampling, as do
-    /// Gemini/DeepSeek. Mirrors the drop in [`to_params`](Self::to_params).
-    pub fn sinks_sampling(&self) -> bool {
-        self.thinking == ThinkingStyle::None || self.sampling_under_thinking
-    }
-
-    /// Just the thinking block (no sampling), with the default effort — the body of the
-    /// [`thinking_params`] wrapper.
-    fn thinking_only(&self, budget: u64) -> Option<Value> {
-        let mut obj = serde_json::Map::new();
-        self.write_thinking(&mut obj, budget, DEFAULT_EFFORT);
-        (!obj.is_empty()).then_some(Value::Object(obj))
-    }
-
-    /// The full `additional_params` blob — thinking (with its effort sink where the
-    /// model has one) plus sampling — or `None` when nothing is set. `effort` is the
-    /// per-role depth lever; it lands where the style takes it (Anthropic adaptive
-    /// → `output_config.effort`; DeepSeek → `reasoning_effort`; the Gemini 3-line
-    /// → `thinkingLevel`), ignored elsewhere.
-    pub fn to_params(
-        &self,
-        budget: u64,
-        temperature: Option<f64>,
-        top_p: Option<f64>,
-        effort: &str,
-    ) -> Option<Value> {
-        let mut obj = serde_json::Map::new();
-        self.write_thinking(&mut obj, budget, effort);
-        let thinking_on = self.thinking != ThinkingStyle::None;
-
-        // Drop sampling when thinking is on and this model won't accept it under
-        // thinking — generalizes the Anthropic case (a custom temperature 400s under
-        // thinking; thinking is the higher-value default, so it wins). DeepSeek/Gemini/
-        // OpenAI accept sampling alongside reasoning, so they keep it.
-        let drop_sampling = thinking_on && !self.sampling_under_thinking;
-        if !drop_sampling {
-            match self.sampling_placement {
-                SamplingPlacement::GeminiGenerationConfig => {
-                    if temperature.is_some() || top_p.is_some() {
-                        let gc = obj
-                            .entry("generationConfig")
-                            .or_insert_with(|| json!({}))
-                            .as_object_mut()
-                            .expect("generationConfig is an object");
-                        if let Some(t) = temperature {
-                            gc.insert("temperature".into(), json!(t));
-                        }
-                        if let Some(p) = top_p {
-                            gc.insert("topP".into(), json!(p));
-                        }
-                    }
-                }
-                SamplingPlacement::TopLevel => {
-                    if let Some(t) = temperature {
-                        obj.insert("temperature".into(), json!(t));
-                    }
-                    if let Some(p) = top_p {
-                        obj.insert("top_p".into(), json!(p));
-                    }
-                }
-            }
-        }
-        (!obj.is_empty()).then_some(Value::Object(obj))
-    }
-
-    /// Write this style's thinking block (and its per-role effort sink) into `obj`.
-    fn write_thinking(&self, obj: &mut serde_json::Map<String, Value>, budget: u64, effort: &str) {
-        match self.thinking {
-            ThinkingStyle::AnthropicBudget => {
-                obj.insert(
-                    "thinking".into(),
-                    json!({ "type": "enabled", "budget_tokens": budget }),
-                );
-            }
-            ThinkingStyle::AnthropicAdaptive => {
-                obj.insert("thinking".into(), json!({ "type": "adaptive" }));
-                // rig 0.34 flattens additional_params into the Messages body; its typed
-                // `output_config` field models only `{format}` and stays `None` (kaibo
-                // sets no output schema), so this flattened key is the only `output_config`
-                // emitted. If kaibo ever adds structured output, revisit — two keys 400.
-                obj.insert("output_config".into(), json!({ "effort": effort }));
-            }
-            ThinkingStyle::GeminiLevel => {
-                // The 3-line's depth lever IS the per-role effort: the values align
-                // ("high"/"low" are valid levels; the default "high" matches
-                // gemini-cli's investigator, rig deserializes it to
-                // ThinkingLevel::High), and like every effort it's a passthrough
-                // string the provider validates. Dropping it for a hardcoded
-                // "high" would silently ignore a slot's `effort = "low"`.
-                obj.insert(
-                    "generationConfig".into(),
-                    json!({ "thinkingConfig": { "thinkingLevel": effort, "includeThoughts": true } }),
-                );
-            }
-            ThinkingStyle::GeminiBudget => {
-                obj.insert(
-                    "generationConfig".into(),
-                    json!({ "thinkingConfig": { "thinkingBudget": budget, "includeThoughts": true } }),
-                );
-            }
-            ThinkingStyle::DeepSeekEffort => {
-                // Explicit-on (the V4 default, but stated so intent survives a default
-                // flip). rig flattens both top-level and round-trips the response
-                // `reasoning_content` so tool loops don't trip DeepSeek's echo-or-400 rule.
-                obj.insert("thinking".into(), json!({ "type": "enabled" }));
-                obj.insert("reasoning_effort".into(), json!(effort));
-            }
-            ThinkingStyle::None => {}
-        }
-    }
-}
-
-/// Per-(provider, model) thinking params, or `None` when the model reasons without a
-/// request toggle. Thin wrapper over [`ModelShape`] using the built-in classifier and
-/// the default effort; the per-phase path with slot overrides goes through
-/// [`Arm::from_slot`].
-pub fn thinking_params(kind: ProviderKind, model: &str, budget: u64) -> Option<Value> {
-    ModelShape::resolve(kind, model, ThinkingStyleOverride::Auto).thinking_only(budget)
-}
-
-/// All of a request's model-shaping params (thinking + sampling) merged into one
-/// `additional_params` blob. Thin wrapper over [`ModelShape::to_params`] using the
-/// built-in classifier and the default effort; the per-phase path with slot
-/// overrides goes through [`Arm::from_slot`].
-pub fn request_params(
-    kind: ProviderKind,
-    model: &str,
-    budget: u64,
-    temperature: Option<f64>,
-    top_p: Option<f64>,
-) -> Option<Value> {
-    ModelShape::resolve(kind, model, ThinkingStyleOverride::Auto).to_params(
-        budget,
-        temperature,
-        top_p,
-        DEFAULT_EFFORT,
-    )
-}
+use super::prompts::{
+    consult_user_prompt, deliberation_prompt, resolve_phase_preamble, ConsultAttachment, Phase,
+    PromptOverrides,
+};
+use super::shaping::{ModelCaps, ModelShape};
 
 // --- The Arm seam ------------------------------------------------------------
 
@@ -910,28 +285,6 @@ impl Arm {
             )
             .await
     }
-}
-
-/// Per-call loop tunables for a phase. Model-tracking knobs (`max_tokens`, the
-/// thinking budget, sampling) ride each [`Arm`] (they track the slot's model);
-/// what remains here are the loop bounds the caller may dial per request, the
-/// sandbox limits, and the progress sink.
-/// One caller-attached file, classified so the driver's prompt can route it to the
-/// right tool. `consult` never inlines an attachment's bytes — text files the model
-/// reads itself with `cat -n`, images it views with `view_image` (the image-analog of
-/// `cat`, present whenever the synth is vision-capable). The server sniffs each file's
-/// magic bytes to set `is_image`, so a `.png` named `.txt` (or vice versa) is routed by
-/// content, not extension — matching how `view_image` re-sniffs authoritatively at read.
-#[derive(Debug, Clone)]
-pub struct ConsultAttachment {
-    /// The path the model passes to `cat -n` or `view_image`: root-relative for a file
-    /// under the project root, the one real tree the consult shell mounts.
-    pub path: String,
-    /// True when the file sniffed as a known image type — route it to `view_image`, not
-    /// the shell. A consult that carries an image attachment must run a vision-capable
-    /// synth (the server gates this up front, the same honest refusal `oneshot`/`batch`
-    /// give a blind model).
-    pub is_image: bool,
 }
 
 #[derive(Debug, Clone)]
@@ -1588,73 +941,6 @@ impl Tool for RunExplore {
     }
 }
 
-/// The `oneshot` preamble: a thin, direct second opinion with no tools and no
-/// codebase access. The caller owns the context, so this never investigates — frame
-/// the model as a capable outside voice answering from what it was handed plus its
-/// own knowledge. Deliberately minimal: no kaish cheatsheet (there are no tools to
-/// drive) and no repo map (oneshot never reads the project).
-pub fn oneshot_preamble() -> String {
-    "You are a capable model giving a direct second opinion to another agent. Answer \
-     the question it sends from the material it provides and your own knowledge — \
-     this call has no codebase access and no tools, so the caller owns the context. \
-     Be precise and useful: reason over exactly what you were handed, and name \
-     explicitly anything you'd need that wasn't given, so the caller can supply it \
-     next turn. Keep your claims grounded in the material and say where its edge is."
-        .to_string()
-}
-
-/// The `batch` preamble: a capable model answering one hard question *offline*, at
-/// max thinking, with no codebase access and no tools. Deliberately **not** a reuse of
-/// [`oneshot_preamble`] — batch is the same toolless shape but a different behavioral
-/// contract, and a cross-model review of the feature caught three places the oneshot
-/// wording misfires for the async lane:
-///
-/// - **No follow-up turn.** A batch item is answered once, offline; the caller cannot
-///   clarify and there is no next turn. oneshot's "name what you'd need rather than
-///   guessing" is right *synchronously* (flagging a gap invites the caller to fill it
-///   next turn) but wrong here — stopping at "I'd need X" burns the caller's one shot
-///   for nothing. The batch contract is *state the assumption, answer under it, flag
-///   what would change* — both the answer and the diagnostic, in one pass.
-/// - **Depth is free.** The lane forces high effort + a generous token floor precisely
-///   because the latency is already accepted. The prompt says so out loud — spend the
-///   room on depth — rather than leaving that intent only in the knobs.
-/// - **Primary answer, not a footnote.** Batch is for asking the best model the hard
-///   question, so the "second opinion" framing under-positions it; the load-bearing
-///   part is "for another agent" (an external advisor owns no context), which we keep.
-///
-/// Positive framing throughout (the CLAUDE.md rule): the old "rather than guessing it"
-/// named the unwanted pathway; the replacement asks for the wanted behavior — a
-/// reasoned, labelled assumption — directly.
-pub fn batch_preamble() -> String {
-    "You are a capable model answering a hard question for another agent, offline. Work \
-     from the material it provides and your own knowledge — this call has no codebase \
-     access and no tools, so the caller owns all the context you have. This is your \
-     single response: there is no follow-up turn and the caller cannot clarify, so make \
-     the answer complete and self-contained, and spend the room you have on depth — \
-     reason the problem all the way through. Be direct and precise. Ground every claim \
-     in the material or your own knowledge, and say where the evidence runs out. Where \
-     something you'd need is missing, state the assumption you're making, answer under \
-     it, and flag what would change if the assumption is wrong."
-        .to_string()
-}
-
-/// Resolve the `batch` phase's system prompt: the operator `[prompts].batch` override
-/// if set, else the built-in [`batch_preamble`]. Batch reads no project (the `oneshot`
-/// shape), so neither the repo map nor house rules splice — the same composition
-/// `oneshot` gets, exposed as a public seam because the batch path lives outside the
-/// `ConsultConfig`-driven loop (it runs on the provider's batch lane, not [`Arm::run`]).
-pub fn batch_system_prompt(override_: Option<&str>) -> String {
-    // Route through the shared `Phase` seam so the `Batch` framing (built-in vs
-    // override, project layers off) is decided in exactly one place — the same one the
-    // resource renders. This path carries a bare override rather than a full
-    // `PromptOverrides`, so wrap it in the one key `Phase::Batch` reads.
-    let prompts = PromptOverrides {
-        batch: override_.map(str::to_string),
-        ..Default::default()
-    };
-    resolve_phase_preamble(Phase::Batch, &prompts, None, None)
-}
-
 /// The `oneshot` seam: one direct completion on the resolved arm — no tools, no
 /// shell, no exploration. The thin counterpart to `consult` (prompt in, answer out)
 /// for when the caller already owns the context and just wants the model's take.
@@ -1722,123 +1008,6 @@ pub async fn oneshot(
         &|| Ok(Vec::new()),
     )
     .await
-}
-
-/// Build the consult driver's user prompt from the question, any caller-supplied
-/// `context`, and any prior session turns. Pure and offline-testable: this framing
-/// is the whole of the context-seed and multi-turn hand-off, so it's worth pinning.
-///
-/// With **no** context and **no** history this is exactly the bare question — a
-/// stateless, unseeded consult is byte-for-byte unchanged. Supplied `context`
-/// (a diff summary, a prior report, pasted source) is framed as *trusted starting
-/// evidence*: a grounded `file:line` rarely needs re-deriving, and the steer is to
-/// investigate for *more* when the context isn't enough — the CLAUDE.md acquisition,
-/// not verification, posture. History prepends the prior `(question, answer)` pairs
-/// and steers the model to re-confirm any span a prior answer cited: the exploration
-/// runs fresh every turn (we never replay the stored report — it'd be stale), so the
-/// code is the ground truth, not the old answer.
-pub fn consult_user_prompt(
-    question: &str,
-    context: Option<&str>,
-    history: &[QaTurn],
-    attached: &[ConsultAttachment],
-) -> String {
-    let context = context.map(str::trim).filter(|c| !c.is_empty());
-    if history.is_empty() && context.is_none() && attached.is_empty() {
-        return question.to_string();
-    }
-    let mut prompt = String::new();
-    if !history.is_empty() {
-        prompt.push_str(
-            "This is a continuing conversation about the same codebase. Earlier turns, \
-             oldest first:\n\n",
-        );
-        for (i, turn) in history.iter().enumerate() {
-            prompt.push_str(&format!(
-                "[Turn {}]\nQ: {}\nA: {}\n\n",
-                i + 1,
-                turn.question,
-                turn.answer
-            ));
-        }
-        prompt.push_str(
-            "Use the earlier turns for context and continuity. Investigate fresh and \
-             re-confirm any `file:line` an earlier answer cited before you rely on it — \
-             the code is the ground truth, not the prior answer.\n\n",
-        );
-    }
-    if let Some(context) = context {
-        prompt.push_str(&format!(
-            "Context the caller supplied (a diff or change summary, a prior report, or \
-             pasted source):\n{context}\n\n\
-             Treat it as trusted starting evidence: when it cites a concrete \
-             `file:line`, trust it rather than re-deriving it. Reach for your tools \
-             when you need more than it gives — a span it references but doesn't quote, \
-             a whole file for the full picture, a detail it left open, or anything the \
-             question reaches that it didn't cover. Where the code you read and the \
-             context genuinely disagree, the code wins.\n\n",
-        ));
-    }
-    if !attached.is_empty() {
-        let (images, texts): (Vec<&ConsultAttachment>, Vec<&ConsultAttachment>) =
-            attached.iter().partition(|a| a.is_image);
-        prompt.push_str(
-            "The caller attached these files as central to the question. They live under \
-             the project root, so a relative path opens directly. Open each as you build \
-             your answer:\n",
-        );
-        if !texts.is_empty() {
-            prompt.push_str("\nText files — read each in full with the shell (`cat -n PATH`):\n");
-            for a in &texts {
-                prompt.push_str(&format!("- {}\n", a.path));
-            }
-        }
-        if !images.is_empty() {
-            // Images are binary — `cat` refuses them; the model has a `view_image` tool
-            // (present because the synth is vision-capable, gated server-side) that hands
-            // it the actual picture. Route images there, never to the shell.
-            prompt.push_str(
-                "\nImages — view each with the `view_image` tool (`view_image PATH`), which \
-                 hands you the picture itself; don't `cat` an image:\n",
-            );
-            for a in &images {
-                prompt.push_str(&format!("- {}\n", a.path));
-            }
-        }
-        prompt.push('\n');
-    }
-    prompt.push_str(&format!("Now answer the current question:\n\n{question}"));
-    prompt
-}
-
-/// The recomposed `consult` driver: one capable model, two tools. Composes the
-/// shared [`kaish_syntax_core`] (for `run_kaish`) and frames `explore` as the way
-/// to cover breadth. Positive framing on purpose — weaker/local models loop on
-/// blanket prohibitions, so reinforce the grounded behavior we want.
-pub fn consult_preamble() -> String {
-    let core = kaish_syntax_core();
-    format!(
-        "You answer a question about a codebase, grounded in evidence and citing \
-         concrete `file:line`. {core}\n\n\
-         You also have a second tool, `explore`: it delegates a broad sweep to a \
-         fast investigator that rips through the repo and reports back with a \
-         curated report — RelevantLocations carrying `file:line`, key symbols, and \
-         snippets. Reach for `explore` to cover breadth — find where a \
-         thing lives, gather the relevant files — and use `run_kaish` to read the \
-         code yourself. When you read directly, read generously in wide passes — a \
-         short file whole with `cat -n FILE`, a big one in spans of a few hundred \
-         lines (`cat -n FILE | sed -n '1,400p'`, then `'401,800p'`) — so each look \
-         lands the code in its context. Build your answer from what \
-         they return: quote the key snippet, name its `file:line`, and let the \
-         evidence carry the claim. Where the evidence settles the question, answer \
-         it fully; where it reaches its edge, say so and name what would close the gap.\n\n\
-         The caller may hand you CONTEXT — a diff or change summary, a prior report, \
-         or pasted source. Treat it as trusted starting evidence: when it cites a \
-         concrete `file:line`, trust it rather than re-deriving it, and spend your \
-         turns getting *more* than it gave — reading a span it left unquoted, a whole \
-         file for the full picture, anything the question reaches past it. Where the \
-         code you read and the context genuinely disagree, the code wins."
-    )
 }
 
 /// Build the recomposed `consult` toolset: `{run_kaish, explore′}`. Factored out so
@@ -1919,28 +1088,6 @@ pub(crate) async fn explore_with(
         &cfg.progress,
     )
     .await
-}
-
-/// Frame a built dossier + the original question into the offline synth's single
-/// user turn — the whole of what `deliberate`'s heavyweight synth reasons over, on
-/// either lane (a batch item's prompt, or the direct lane's one local completion).
-/// Pure, so the wire shape is pinned without a network.
-///
-/// The framing installs the deliberate posture: the dossier is *trusted* investigated
-/// evidence (a fast explorer read the real spans and cited them), so the synth spends
-/// its one offline turn reasoning the question all the way through, not re-verifying
-/// cites it can't cheaply re-derive — and names the edge of the evidence rather than
-/// guessing past it (the "thin dossier deliberating on air" failure the spec warns of).
-pub fn deliberation_prompt(question: &str, dossier: &str) -> String {
-    format!(
-        "A fast explorer investigated this codebase READ-ONLY and assembled the dossier \
-         below — spans it read from the real, current source, cited by `file:line`. Trust \
-         those citations as accurate and deliberate deeply over the question using this \
-         evidence: reason it all the way through, and say where the evidence runs out. If \
-         the dossier leaves a load-bearing detail open, reason under a stated assumption \
-         and flag what would change if it's wrong.\n\n\
-         ## Question\n{question}\n\n## Dossier\n{dossier}"
-    )
 }
 
 /// Run `deliberate`'s offline synth on the **direct** lane: one long, toolless local
@@ -2070,6 +1217,7 @@ pub async fn consult(
 
 #[cfg(test)]
 mod tests {
+    use super::super::shaping::thinking_params;
     use super::*;
     use std::time::Duration;
 
@@ -2788,33 +1936,6 @@ mod tests {
         );
     }
 
-    /// The deliberation prompt is the whole context the offline synth reasons over, so
-    /// pin its shape: both the question and the dossier survive, the question is read
-    /// before the evidence, and the trust-the-cites posture is installed (the guard
-    /// against a synth burning its one turn re-verifying, or deliberating on air).
-    #[test]
-    fn deliberation_prompt_carries_question_then_dossier_and_frames_trust() {
-        let p = deliberation_prompt("Is the retry path safe?", "src/retry.rs:12 fn retry()");
-        assert!(
-            p.contains("Is the retry path safe?"),
-            "question present: {p}"
-        );
-        assert!(
-            p.contains("src/retry.rs:12 fn retry()"),
-            "dossier present: {p}"
-        );
-        let q = p.find("## Question").expect("has a Question section");
-        let d = p.find("## Dossier").expect("has a Dossier section");
-        assert!(
-            q < d,
-            "the question is framed before the dossier evidence: {p}"
-        );
-        assert!(
-            p.contains("Trust") && p.contains("evidence"),
-            "installs the trusted-evidence posture: {p}"
-        );
-    }
-
     /// The `direct` lane: `deliberate_direct` runs the offline synth as ONE toolless
     /// turn over the dossier and returns its deliberation. Teeth: the synth must be
     /// toolless (no `run_kaish`/`explore` — it reasons over the handed dossier, it
@@ -3450,30 +2571,6 @@ mod tests {
         );
     }
 
-    /// The Gemini 3-line's depth lever IS the per-role effort: a slot's `effort`
-    /// must land as `thinkingLevel` (the values align — "high"/"low" are valid
-    /// levels), not be silently dropped into a hardcoded max-depth. A user
-    /// setting `effort = "low"` on a gemini-3 slot means it.
-    #[test]
-    fn gemini_level_takes_the_per_role_effort_as_its_thinking_level() {
-        let shape = ModelShape::resolve(
-            ProviderKind::Gemini,
-            "gemini-3-pro-preview",
-            ThinkingStyleOverride::Auto,
-        );
-        let params = shape.to_params(8192, None, None, "low").unwrap();
-        assert_eq!(
-            params["generationConfig"]["thinkingConfig"]["thinkingLevel"], "low",
-            "the slot's effort is the level"
-        );
-        // The default effort keeps today's wire shape byte-for-byte.
-        let params = shape.to_params(8192, None, None, DEFAULT_EFFORT).unwrap();
-        assert_eq!(
-            params["generationConfig"]["thinkingConfig"]["thinkingLevel"],
-            "high"
-        );
-    }
-
     /// `Arm::from_slot` is the single live construction point, and per-call
     /// overrides build slots that never saw config load's budget validation —
     /// so the thinking_budget < max_tokens rule must hold here too, as the same
@@ -3742,144 +2839,6 @@ mod tests {
         assert!(
             !names.iter().any(|n| n == "view_image"),
             "a blind synth must not get view_image, got {names:?}"
-        );
-    }
-
-    /// The explorer preamble carries the behaviors we measured into it — the
-    /// whole-file reading directive (the lite-explorer win, 48→23 turns), the
-    /// context-buffer `grep` idiom, and the three report sections the synth side now
-    /// expects. Pure and offline; pins the prose so a future edit can't silently
-    /// drop any of it (the synth preambles are written against this shape).
-    #[test]
-    fn report_preamble_keeps_the_reading_directive_and_report_shape() {
-        let p = report_preamble();
-        // Reading strategy: whole (short) files, wide spans for big ones, grep buffer.
-        assert!(p.contains("cat -n FILE"), "whole-file read idiom: {p}");
-        assert!(
-            p.to_lowercase().contains("whole"),
-            "the whole-file directive must survive: {p}"
-        );
-        assert!(
-            p.contains("sed -n '1,400p'"),
-            "big-file wide-span idiom: {p}"
-        );
-        assert!(
-            p.contains("grep -rn -B4 -A8"),
-            "grep context-buffer idiom: {p}"
-        );
-        // The report template the consult driver preamble is written
-        // against — keep the three section names in lockstep with those.
-        for section in ["SummaryOfFindings", "RelevantLocations", "ExplorationTrace"] {
-            assert!(p.contains(section), "missing report section {section}: {p}");
-        }
-    }
-
-    /// The batch preamble encodes the async lane's distinct contract — the three things
-    /// a cross-model review flagged the oneshot wording getting wrong for batch. These
-    /// are behavioral promises, so they get a test that fails if the prose drifts back
-    /// toward the synchronous oneshot framing.
-    #[test]
-    fn batch_preamble_encodes_the_offline_one_shot_contract() {
-        let p = batch_preamble();
-        let lower = p.to_lowercase();
-        // (1) No follow-up turn — be complete and self-contained in one response.
-        assert!(
-            lower.contains("single response") && lower.contains("no follow-up"),
-            "batch must tell the model it gets exactly one offline response: {p}"
-        );
-        // (2) Depth is free here — spend the budget the lane forces on.
-        assert!(lower.contains("depth"), "batch must ask for depth: {p}");
-        // (3) Assume-and-answer, not flag-and-stall: state the assumption and answer
-        // under it (the synchronous oneshot would say "name what you'd need").
-        assert!(
-            lower.contains("assumption") && lower.contains("answer under it"),
-            "batch must steer toward assume-and-answer, not flag-and-stall: {p}"
-        );
-        // Positive framing (the CLAUDE.md rule): it must not reintroduce the negative
-        // "rather than guessing" pathway the oneshot line used.
-        assert!(
-            !lower.contains("guess"),
-            "batch preamble must stay positively framed — no 'guess' pathway: {p}"
-        );
-        // Still the toolless, contextless shape it shares with oneshot.
-        assert!(
-            lower.contains("no codebase access") && lower.contains("no tools"),
-            "batch is the toolless, contextless shape: {p}"
-        );
-    }
-
-    /// `[prompts].batch` fully replaces the built-in batch preamble; absent, the
-    /// built-in stands. Batch reads no project, so nothing else splices.
-    #[test]
-    fn batch_system_prompt_honors_the_override() {
-        assert_eq!(batch_system_prompt(None), batch_preamble());
-        assert_eq!(
-            batch_system_prompt(Some("custom batch frame")),
-            "custom batch frame"
-        );
-    }
-
-    /// The single `Phase` seam both the tools and the `kaibo://prompts` resource go
-    /// through: each phase resolves to its own built-in default, an override wins per
-    /// key, and the `[orientation]`/`[context]` project layers splice *only* for the
-    /// project-reading phases (explorer + consult) — never for the caller-owns-context
-    /// phases (oneshot + the offline batch synth), even when the layers are passed.
-    #[test]
-    fn resolve_phase_preamble_routes_each_phase_and_gates_project_layers() {
-        let base = PromptOverrides::default();
-        assert_eq!(
-            resolve_phase_preamble(Phase::Explorer, &base, None, None),
-            report_preamble()
-        );
-        assert_eq!(
-            resolve_phase_preamble(Phase::Consult, &base, None, None),
-            consult_preamble()
-        );
-        assert_eq!(
-            resolve_phase_preamble(Phase::Oneshot, &base, None, None),
-            oneshot_preamble()
-        );
-        assert_eq!(
-            resolve_phase_preamble(Phase::Batch, &base, None, None),
-            batch_preamble()
-        );
-
-        let map = "PROJECT FILES.\nsrc/lib.rs";
-        let rules = "operator house rule";
-        // The reading phases splice both project layers.
-        for phase in [Phase::Explorer, Phase::Consult] {
-            let p = resolve_phase_preamble(phase, &base, Some(map), Some(rules));
-            assert!(
-                p.contains(map) && p.contains(rules),
-                "{} must splice the project layers",
-                phase.label()
-            );
-            assert!(phase.reads_project());
-        }
-        // The context-owning phases drop them even when passed.
-        for phase in [Phase::Oneshot, Phase::Batch] {
-            assert_eq!(
-                resolve_phase_preamble(phase, &base, Some(map), Some(rules)),
-                resolve_phase_preamble(phase, &base, None, None),
-                "{} must ignore the project layers",
-                phase.label()
-            );
-            assert!(!phase.reads_project());
-        }
-
-        // An override wins over the built-in, per key.
-        let ov = PromptOverrides {
-            consult: Some("CUSTOM DRIVER".into()),
-            ..Default::default()
-        };
-        assert_eq!(
-            resolve_phase_preamble(Phase::Consult, &ov, None, None),
-            "CUSTOM DRIVER"
-        );
-        // ...and doesn't bleed into a sibling phase.
-        assert_eq!(
-            resolve_phase_preamble(Phase::Oneshot, &ov, None, None),
-            oneshot_preamble()
         );
     }
 
@@ -4194,132 +3153,6 @@ mod tests {
             count_model_turns(&history),
             2,
             "only the two assistant messages count; the inserted image turn does not"
-        );
-    }
-
-    /// No session history ⇒ the prompt is *exactly* the bare question. This pins the
-    /// promise that a stateless consult is byte-for-byte its pre-session behavior.
-    #[test]
-    fn empty_history_yields_the_bare_question() {
-        assert_eq!(
-            consult_user_prompt("Where is the sandbox enforced?", None, &[], &[]),
-            "Where is the sandbox enforced?"
-        );
-    }
-
-    /// Text attachments are named in the prompt (for the model to `cat` itself) and
-    /// steered to read them in full — and they're listed before the question, like context.
-    #[test]
-    fn attached_files_are_named_for_the_model_to_read() {
-        let prompt = consult_user_prompt(
-            "Does the diff weaken the sandbox?",
-            None,
-            &[],
-            &[text_attach("changes.diff"), text_attach("src/sandbox.rs")],
-        );
-        assert!(
-            prompt.contains("changes.diff"),
-            "names each attached file:\n{prompt}"
-        );
-        assert!(
-            prompt.contains("src/sandbox.rs"),
-            "names each attached file:\n{prompt}"
-        );
-        assert!(
-            prompt.contains("cat -n"),
-            "steers the model to read them with the shell:\n{prompt}"
-        );
-        let listed = prompt.find("changes.diff").unwrap();
-        let question = prompt.find("Does the diff weaken").unwrap();
-        assert!(
-            listed < question,
-            "attachments precede the question:\n{prompt}"
-        );
-    }
-
-    /// An image attachment must be routed to `view_image`, never to `cat` (which refuses
-    /// binary). With a mix, each file lands under the right instruction: text → `cat -n`,
-    /// image → `view_image`. This is the prompt half of the image-attach support; the
-    /// server gates a vision-blind synth before we ever get here.
-    #[test]
-    fn image_attachments_are_routed_to_view_image_not_cat() {
-        let prompt = consult_user_prompt(
-            "What does the banner show, and does it match the brand doc?",
-            None,
-            &[],
-            &[
-                image_attach("docs/brand/banner-teal.png"),
-                text_attach("docs/brand/README.md"),
-            ],
-        );
-        // The image is steered to view_image and explicitly kept away from the shell.
-        let view_at = prompt
-            .find("view_image")
-            .expect("image must be routed to view_image");
-        let img_at = prompt
-            .find("banner-teal.png")
-            .expect("image is named in the prompt");
-        // The image name sits under the view_image instruction, not the cat -n one.
-        let cat_at = prompt.find("cat -n").expect("text section names cat -n");
-        assert!(
-            view_at < img_at,
-            "the image is listed under the view_image instruction:\n{prompt}"
-        );
-        // The text file is under the cat -n section.
-        let readme_at = prompt
-            .find("README.md")
-            .expect("text file is named in the prompt");
-        assert!(
-            cat_at < readme_at,
-            "the text file is listed under the cat -n instruction:\n{prompt}"
-        );
-    }
-
-    fn text_attach(path: &str) -> ConsultAttachment {
-        ConsultAttachment {
-            path: path.to_string(),
-            is_image: false,
-        }
-    }
-
-    fn image_attach(path: &str) -> ConsultAttachment {
-        ConsultAttachment {
-            path: path.to_string(),
-            is_image: true,
-        }
-    }
-
-    /// With history, every prior turn appears, the current question appears, and the
-    /// turns precede the current question (the model reads context before the ask).
-    #[test]
-    fn history_is_replayed_before_the_current_question_in_order() {
-        let history = vec![
-            QaTurn::new("What is kaish?", "A read-only shell (src/sandbox.rs)."),
-            QaTurn::new("Who calls it?", "consult drives it (src/consult.rs)."),
-        ];
-        let prompt = consult_user_prompt("And explore?", None, &history, &[]);
-
-        for needle in [
-            "What is kaish?",
-            "A read-only shell (src/sandbox.rs).",
-            "Who calls it?",
-            "consult drives it (src/consult.rs).",
-            "And explore?",
-        ] {
-            assert!(
-                prompt.contains(needle),
-                "prompt must carry {needle:?}:\n{prompt}"
-            );
-        }
-        // Ordering: the first prior turn comes before the second, and both come
-        // before the current question.
-        let first = prompt.find("What is kaish?").unwrap();
-        let second = prompt.find("Who calls it?").unwrap();
-        let current = prompt.find("And explore?").unwrap();
-        assert!(first < second, "turns must be oldest-first");
-        assert!(
-            second < current,
-            "history must precede the current question"
         );
     }
 }
